Restructure PHP files.
